backend/pyproject.toml
[project]
name = "nlq-helpdesk"
version = "1.0.0"
description = "NLQ Helpdesk - Gen AI conversational helpdesk with RAG"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "google-genai>=1.0.0",
    "chromadb-client>=1.0.0",
    "aiosqlite>=0.20.0",
    "python-multipart>=0.0.18",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
